fix sorting to show unit No 1 at start
